fix(linting error): convert findRoute to hook, useRouteFinder, move to separate utils file
