+++ src/components/VersionSwitcher/index.tsx
@@ -1,34 +1,25 @@
 import { Flex } from '@aws-amplify/ui-react';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
-import flatDirectory from 'src/directory/flatDirectory.json';
 import { IconCheck } from '@/components/Icons';
 import { PLATFORM_VERSIONS } from '@/data/platforms';
 import classNames from 'classnames';
 import { trackVersionChange } from '@/utils/track';
+import { useRouteFinder } from '@/utils/useRouteFinder';
 import { BUILD_A_BACKEND, PREV_BUILD_A_BACKEND } from '@/data/routes';
-
-const findRoute = (platform, isPrev) => {
-  const router = useRouter();
-  const path = router.pathname;
-  const newRoute = isPrev
-    ? path.replace('/[platform]/prev', '/[platform]')
-    : path.replace('/[platform]', '/[platform]/prev');
-  const pageNode = flatDirectory[newRoute];
-  if (pageNode && pageNode.platforms && pageNode.platforms.includes(platform)) {
-    return newRoute;
-  }
-};
 
 export const VersionSwitcher = ({ platform, isPrev, ...rest }) => {
   const router = useRouter();
   const pathname = router.pathname;
   const versions = PLATFORM_VERSIONS[platform];
-  const switchPath = findRoute(platform, isPrev);
+  const switchPath = useRouteFinder(platform, isPrev);
   let path = isPrev ? BUILD_A_BACKEND : PREV_BUILD_A_BACKEND;
-  if (switchPath &&
-    (pathname.startsWith(BUILD_A_BACKEND) || pathname.startsWith(PREV_BUILD_A_BACKEND))
-  ) path = switchPath;
+  if (
+    switchPath &&
+    (pathname.startsWith(BUILD_A_BACKEND) ||
+      pathname.startsWith(PREV_BUILD_A_BACKEND))
+  )
+    path = switchPath;
 
   const inactiveHref = {
     pathname: path,
